requirements.txt
SpeechRecognition
textual
rich
llama-cpp-python
google-genai
requests
PyGithub
Pillow
PyYAML
tika
PyAudio
